requirements.txt
streamlit>=1.22.0
pinecone-client>=2.2.0
pandas>=1.4.0
sentence-transformers>=2.2.2
python-dotenv>=0.20.0
tqdm>=4.64.0 